Move file and prefer methods per feedback.
H/t @jeremiak:
* The PageCounter component isn't tied to policies, so don't place it in a
  sub-folder.
* We don't use getters anywhere else, so let's not introduce them in the
  policy page.
