+++ ui/components/page-counter.js
